+++ src/main/java/lotterymachine/LotteryMachineApplication.java
@@ -1,8 +1,8 @@
 package lotterymachine;
 
 import lotterymachine.domain.*;
-import lotterymachine.domain.vo.Count;
-import lotterymachine.domain.vo.Money;
+import lotterymachine.domain.Count;
+import lotterymachine.domain.Money;
 import lotterymachine.view.InputView;
 import lotterymachine.view.OutputView;
 
@@ -12,7 +12,7 @@ import java.util.List;
 public class LotteryMachineApplication {
     public static void main(String[] args) {
         Money money = new Money(InputView.getAmount());
-        Count count = new Count(InputView.getPassivityPurchaseCount(), money.getPurchasePossibleCount());
+        Count count = createCount(money);
         LotteryTickets lotteryTickets = createLotteryTickets(count);
         OutputView.printLotteryPurchaseCount(count);
         OutputView.printLotteryTickets(lotteryTickets);
@@ -22,6 +22,11 @@ public class LotteryMachineApplication {
         WinningResult winningResult = WinningResult.create(lotteryTickets, winningLottery);
         OutputView.printWinningLotteryResults(winningResult.getResult());
         OutputView.printProfitRate(Yield.of(winningResult.findTotalProfit(), money.getValue()).getProfitRate());
+    }
+
+    private static Count createCount(Money money) {
+        int passivityPurchaseCount = InputView.getPassivityPurchaseCount();
+        return new Count(passivityPurchaseCount, money.getPurchasePossibleCount() - passivityPurchaseCount, money.getPurchasePossibleCount());
     }
 
     private static LotteryTickets createLotteryTickets(Count count) {
