+++ src/main/java/lotterymachine/view/OutputView.java
@@ -3,7 +3,7 @@ package lotterymachine.view;
 import lotterymachine.domain.LotteryTicket;
 import lotterymachine.domain.LotteryTickets;
 import lotterymachine.domain.WinningLotteryRank;
-import lotterymachine.domain.vo.Count;
+import lotterymachine.domain.Count;
 
 import java.util.Map;
